implement center title and show hide title patch
inspired by centertitle and notitle patch, but make it toggleable
and change the center calculation, original centertitle decide that
the center position is the center of bar width.
this patch, decide that center is calculate from the available space
for the title, which is (bar width - tag and layout area - status area).