update config: update rules
a rule that base on WINTYPE without specified class must be placed at
the very end of the rules array to ensure the rule matching worked as
expected.
e.g. firefox and it's PictureInPicture window (have WINTYPE of UTILITY),
if we define a rule with firefox as it class and set it to non floating,
and there are a rule for WINTYPE UTILITY that set to be floating before
the rule for firefox class, the firefox PictureInPicture window will not
be floated because they follow the rule for it class, i.e. firefox.