[ sendmon same tags ] Initial implementation
By default, sendmon function will assign a new tagset to the selected
client, so that the selected client will be in the currently viewed tag
on the next/previous monitor.

This patch causes the sendmon function to send the selected client to
the same tag on the next/previous monitor.